OrgSegNet/requirements.txt
onnx
onnxruntime-gpu
opencv-python==4.5.5.62
Shapely
scikit-image
scikit-learn
scipy
pandas
networkx
pickleshare
numpy
Pillow
matplotlib
mmcv
mmcv-full